updating readme to include TDL improvements
after revisiting this tool, i am amazed that i was satisfied with this. To prevent developers from having to change, the parser must accommodate as much as possible. In that regard, this solution still has much to work on. It is very difficult to find a high-quality example Visio that is unideal in some way, such that my test cases are all failures and my parsable alternatives are unpresentable. This needs work.